+++ data/blog/how-to-share-assets-to-multiple-appslibraries-in-nx.mdx
@@ -2,7 +2,7 @@
 title: How to share assets to multiple apps/libraries in Nx
 date: '2022-05-24'
 tags: ['Nx', 'Workspaces', 'Monorepo', 'Builds']
-draft: true
+draft: false
 summary: A guide on how to share assets to all your apps/libraries in a Nx workspace
 layout: PostLayout
 author: 'Dale Larroder'
