+++ package.json
@@ -24,7 +24,7 @@
     "@vercel/analytics": "0.1.1",
     "classnames": "2.3.2",
     "cmdk": "0.1.20",
-    "contentlayer": "0.2.9",
+    "contentlayer": "0.3.2",
     "esbuild": "0.16.12",
     "framer-motion": "10.12.16",
     "github-slugger": "1.4.0",
@@ -32,7 +32,7 @@
     "logrocket": "3.0.1",
     "mdx-bundler": "9.0.1",
     "next": "^13.4.6",
-    "next-contentlayer": "0.2.9",
+    "next-contentlayer": "0.3.2",
     "next-themes": "0.2.1",
     "nprogress": "0.2.0",
     "react": "^18.2.0",
