--- pages/blog/page/[page].tsx
@@ -1,63 +1,0 @@
-import { PageSEO } from '@/components/SEO';
-import siteMetadata from '@/data/siteMetadata';
-import ListLayout from '@/layouts/MDX/ListLayout';
-import MainLayout from '@/layouts/MainLayout';
-import { allCoreContent, sortedBlogPost } from '@/lib/utils/contentlayer';
-import { POSTS_PER_PAGE } from '@/types/default';
-import { allBlogs } from 'contentlayer/generated';
-import { InferGetStaticPropsType } from 'next';
-
-export const getStaticPaths = async () => {
-  const totalPosts = sortedBlogPost(allBlogs);
-  const totalPages = Math.ceil(totalPosts.length / POSTS_PER_PAGE);
-  const paths = Array.from({ length: totalPages }, (_, i) => ({
-    params: { page: (i + 1).toString() },
-  }));
-
-  return {
-    paths,
-    fallback: false,
-  };
-};
-
-export const getStaticProps = async (context: { params: { page: any } }) => {
-  const {
-    params: { page },
-  } = context;
-  const posts = sortedBlogPost(allBlogs);
-  const pageNumber = parseInt(page as string);
-  const initialDisplayPosts = posts.slice(
-    POSTS_PER_PAGE * (pageNumber - 1),
-    POSTS_PER_PAGE * pageNumber
-  );
-  const pagination = {
-    currentPage: pageNumber,
-    totalPages: Math.ceil(posts.length / POSTS_PER_PAGE),
-  };
-
-  return {
-    props: {
-      initialDisplayPosts: allCoreContent(initialDisplayPosts),
-      posts: allCoreContent(posts),
-      pagination,
-    },
-  };
-};
-
-export default function PostPage({
-  posts,
-  initialDisplayPosts,
-  pagination,
-}: InferGetStaticPropsType<typeof getStaticProps>) {
-  return (
-    <MainLayout>
-      <PageSEO title={siteMetadata.title} description={siteMetadata.description} />
-      <ListLayout
-        posts={posts}
-        initialDisplayPosts={initialDisplayPosts}
-        pagination={pagination}
-        title="All Posts"
-      />
-    </MainLayout>
-  );
-}
